+++ 10-file-system/02-my-rm-dir.js
@@ -1,7 +1,11 @@
 #!/usr/bin/node
 
-var fs = require('fs');
+const fs  = require('fs'),
+      dir = process.argv[2];
 
-var dir = process.argv[2];
-
-if(fs.statSync(dir).isDirectory())  fs.rmdirSync(dir);
+if(fs.existsSync(dir)) {
+  if(fs.statSync(dir).isDirectory())  fs.rmdirSync(dir);
+} else {
+  console.error('%s not exist!', dir);
+  process.exit(1);
+}
