+++ 10-file-system/06-my-rm.js
@@ -1,9 +1,18 @@
 #!/usr/bin/node
 
-var fs = require('fs');
-var join = require('path').join;
+const fs   = require('fs'),
+      join = require('path').join,
+      src  = process.argv[2];
 
-var src = process.argv[2];
+if(typeof(src) === 'undefined') {
+  console.error('请指定要删除的文件名或者目录名！');
+  process.exit(1);
+}
+
+if(!fs.existsSync(src)) {
+  console.error('%s not exist!', src);
+  process.exit(2);
+}
 
 if(fs.statSync(src).isFile())  fs.unlinkSync(src);
